Swipe delete without reloading all of the cards and add some comments and other cleanup
We can use adapter.notifyItemRemoved and remove it from the cardList in adapater directly since we do understand the position in this case. It works better because it does not require reloading everything and causes less bugs with the view model.

I am not sure if this was the proper way to do this. If not I can revert the change to this method

It also removes the need for a callback in DownloadButtonSetup