Dokka: Fix conflicts between library platforms if the same class exists twice.
Also tags classes, methods, properties, etc., from the library and app as Android, Common, and Kotlin. These names can be changed if desired by setting the display name in the source sets.

Meant to fix:
```
w: [:docs:dokkaGeneratePublicationHtml] For WebViewResolver: expected 1 page, but got 2
w: [:docs:dokkaGeneratePublicationHtml] For Log: expected 1 page, but got 2
```

As well as a bunch of "Failed to download package-list" by explicitly setting which ones we need.