Add jdk toolchain
For better support for building with later java versions which have deprecated compiling with Java 8 so it doesn't always give deprecation warnings.

Starting with AGP 9, building with Java 17 is no longer supported, Java 8 still works for compiling, but building will require Java 21 or later so this is preparing for that. It might still be out a bit but it's probably a good thing to start preparing for it.

This PR should have no affect to the current build process other than making it easier and more clear to build with more modern Java versions without loosing any legacy support or being spammed with a bunch of warnings.