Use JDK 21 in workflows
Also does some minor cleanup to workflows.

Starting with AGP 9, building with Java 17 is no longer supported, Java 8 still works for compiling, but building will require Java 21 or later so this is preparing for that. It might still be out a bit but it's probably a good thing to start preparing for it and using it here in workflows, etc... would also help to encourage extension developers and other developers who contribute to the app to use it more for a more gradual migration to it, akin to when we deprecate something so it doesn't take developers by surprise for those that still use versions lower than Java 21 to build.