Improve caching system for actions
This requires adding a new secret in GitHub settings, `GRADLE_ENCRYPTION_KEY` which can be generated with `openssl rand -base64 16`.

This significantly improves caching performance, not only supporting build cache, but configuration cache as well. The encryption key is necessary to enable that for security. It won't work without it. Using this action also provides more information on build status in workflow results page.