Remove "Update to prereleases" setting
This setting is very confusing on stable, when there is a seperate flavor for prerelease that is needed. Also causes other issues and conflicts with what you are updating on stable, and is annoying to always get update dialogs on debug versions when it doesn't work or do anything in debug version.