Use assemblePrereleaseRelease for building prerelease
It's more explicit that way.

This also adds missing `MDL_API_KEY` to archive build, which seems was missed.

Additionally, this also removes running `build` manually, it's unnecessary now. In the future, we can also probably remove `androidSourcesJar` and `sources.jar` from releases, every since the library split it really hasnt been useful as we don't even include library sources in it, so we don't actually need it. But for now, we can leave it. I also plans to have `makeJar` auto run with the release build type to auto include for all releases, including stable so we can have `classes.jar` for stable releases as well (though it would still need included in release artifacts still). We can also create a new workflow triggered by `workflow_dispatch` for manually triggering to streamline creating stable releases and auto include this as well. Those are just some things I'm planning for the future (if even wanted), but this is just a start for that.